Issue #8: SMPP - Message Id: Hexadecimal vs Decimal.
Add a new property in the SmppConfiguration called dlrIdConversion that specifies the following values: No conversion, Hexa to Decimal and Decimal to Hexa.
Modify the SmppConnector to leave the submit response message id intact and convert the id that comes in the DLR (without modifying it).
Add some test to validate the issue.
Change the smpp-server to version 0.1.1 that allows to return hexadecimal values in the message id.