Issue #11: Acceptors should be able to be defined inside other acceptors.
AbstractConfiguration should recieve an acceptor as the value of a property, mapProperty or listProperty.
Change the connectors schema to allow an acceptor inside a property, mapProperty or listProperty.
Create an AndAcceptor that receives a collection of acceptors.
Add the logic in AbstractConfiguration to allow an 'and' tag.
Change the connectors schema to allow the 'and' tag inside the 'acceptors' tag.